+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Version 1.71
+* issue 7165: add method to get contact id with study participant id
+
 ## Version 1.70
-* issue 7165: add method to get volunteer contact ID with study participant id
+* issue 7165: add method to check if volunteer is withdrawn
 * issue 7460: add parameter mode to caleligibility job so it can also recalculate for status invited/invitation pending/accepted